// src/schema/ticket/tests/singleEventTickets.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type SingleEventTicketsQueryVariables = Types.Exact<{
  id: Types.Scalars['String']['input'];
}>;


export type SingleEventTicketsQuery = { __typename?: 'Query', event: { __typename?: 'Event', id: string, name: string, tickets: Array<{ __typename?: 'Ticket', id: string, name: string, quantity: number | null }> } | null };


export const SingleEventTickets = gql`
    query SingleEventTickets($id: String!) {
  event(id: $id) {
    id
    name
    tickets {
      id
      name
      quantity
    }
  }
}
    `;